calls the ListStudent component
