+++ client/src/App.jsx
@@ -1,7 +1,8 @@
 import './App.css'
 import 'bootstrap/dist/css/bootstrap.min.css';
 import MyNavBar from './components/Navbar'
-import ListStudents from './components/ListStudents'
+import ListStudents from './components/ListContacts'
+import ListContacts from './components/ListContacts';
 
 
 function App() {
@@ -9,7 +10,7 @@ function App() {
   return (
     <div className="App">
       <MyNavBar />
-      <ListStudents />
+      <ListContacts />
 
     </div>
   )
